% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddR.R
\name{available_backends}
\alias{available_backends}
\title{List all known supported backends}
\usage{
available_backends()
}
\value{
character vector
}
\description{
List all known supported backends
}
\seealso{
\code{\link{useBackend}}
}

