% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\name{rbind}
\alias{cbind}
\alias{rbind}
\title{rbindddR}
\usage{
rbind(..., deparse.level = 1)

cbind(..., deparse.level = 1)
}
\arguments{
\item{...}{objects to rbind or cbind}

\item{deparse.level}{Does nothing so far.}
}
\value{
bound (cbind or rbind) dobject
}
\description{
rbindddR

cbindddR
}

