\name{bpm_time}
\alias{bpm_time}
\title{get the time in seconds for a given bpm / count}
\usage{
  bpm_time(bpm = 120, count = 1)
}
\arguments{
  \item{bpm}{beats per minute desired}

  \item{count}{"thirtytwo" = 32 bars (or simply, 32 as
  numeric) "thirtytwo_" = 1/32nd note, or simpy 1/32 as
  numeric}
}
\value{
  time(in seconds) of a given note
}
\description{
  get the time in seconds for a given bpm / count
}
\examples{
#not run
}

