% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{$,DObject-method}
\alias{$,DObject-method}
\title{Extracts elements of a distributed object matching the name.}
\usage{
\S4method{$}{DObject}(x, name)
}
\arguments{
\item{x}{The object to get the named element from.}

\item{name}{The name vector to retrieve elements with.}
}
\description{
Extracts elements of a distributed object matching the name.
}

