% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{cbind,DObject-method}
\alias{cbind,DObject-method}
\title{Column binds the objects.}
\usage{
\S4method{cbind}{DObject}(..., deparse.level = 1)
}
\arguments{
\item{...}{Objects to column bind.}

\item{deparse.level}{Does nothing so far.}
}
\value{
A dobject with the operands (and their partitions) cbinded.
}
\description{
Column binds the objects.
}

