% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{colMeans,DObject-method}
\alias{colMeans,DObject-method}
\title{Gets the column means for a distributed array or data.frame.}
\usage{
\S4method{colMeans}{DObject}(x, na.rm = FALSE, dims = 1L)
}
\arguments{
\item{x}{The object to get the column means from.}

\item{na.rm}{If TRUE, will remove NAs.}

\item{dims}{Currently does nothing.}
}
\description{
Gets the column means for a distributed array or data.frame.
}

