% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{colnames,DObject-method}
\alias{colnames,DObject-method}
\title{Gets the colnames for the distributed object.}
\usage{
\S4method{colnames}{DObject}(x)
}
\arguments{
\item{x}{The distributed object to get the colnames for.}
}
\description{
Gets the colnames for the distributed object.
}

