% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{colSums,DObject-method}
\alias{colSums,DObject-method}
\title{Get the column sums for a distributed array or data.frame.}
\usage{
\S4method{colSums}{DObject}(x, na.rm = FALSE, dims = 1L)
}
\arguments{
\item{x}{The object to get the column sums from.}

\item{na.rm}{If TRUE, will remove NAs.}

\item{dims}{Currently does nothing.}
}
\description{
Get the column sums for a distributed array or data.frame.
}

