\name{ddr_init}
\alias{ddr_init}
\title{initialize wave player and attaches note variables}
\usage{
  ddr_init(player = "/Applications/'QuickTime Player.app'/Contents/MacOS/'QuickTime Player'",
    shut_up = TRUE)
}
\arguments{
  \item{path_to_player}{the path to your wave player,
  defaults to Mac OSX settings.}
}
\value{
  initializes the wave player, attaches note variables
}
\description{
  initialize wave player and attaches note variables
}
\examples{
#not run
}

