% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{dimnames,DObject-method}
\alias{dimnames,DObject-method}
\title{Gets the dimnames for the distributed object.}
\usage{
\S4method{dimnames}{DObject}(x)
}
\arguments{
\item{x}{The distributed object to get the dimnames for.}
}
\description{
Gets the dimnames for the distributed object.
}

