% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddR.R, R/pobject.R
\docType{methods}
\name{do_collect}
\alias{do_collect}
\alias{do_collect,ParallelObj,integer-method}
\title{Backend implemented function to move data from storage to the calling
context (node).}
\usage{
do_collect(x, parts)

\S4method{do_collect}{ParallelObj,integer}(x, parts)
}
\arguments{
\item{x}{The distributed object to fetch data from.}

\item{parts}{The parts (indices) of the distributed object to fetch.}
}
\value{
The data returned as a list, matrix, or data.frame.
}
\description{
Backend implemented function to move data from storage to the calling
context (node).
}

