% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{[}
\alias{[}
\alias{[,DObject-method}
\title{Extract parts of a distributed object.}
\usage{
\S4method{[}{DObject}(x, i, j,...,drop=TRUE)
}
\arguments{
\item{x}{The distributed object to get the parts of.}

\item{i}{The row index or indices to extract with.}

\item{j}{The column index or indices to extract with.}

\item{...}{Other args.}

\item{drop}{If TRUE, vectorizable results will become vectors.}
}
\description{
Extract parts of a distributed object.
}

