% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddR.R, R/pobject.R
\docType{methods}
\name{get_parts}
\alias{get_parts}
\alias{get_parts,ParallelObj,integer-method}
\alias{get_parts,ParallelObj,missing-method}
\title{Gets the partitions to a distributed object, given an index.}
\usage{
get_parts(x, index, ...)

\S4method{get_parts}{ParallelObj,missing}(x, index, ...)

\S4method{get_parts}{ParallelObj,integer}(x, index, ...)
}
\arguments{
\item{x}{The distributed object to dispatch on.}

\item{index}{The index or indices of the partitions to fetch.}

\item{...}{Other options (not in use currently)}
}
\value{
A list containing the partitions of 'x'.
}
\description{
Gets the partitions to a distributed object, given an index.
}

