% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddR.R
\name{getBestOutputPartitioning}
\alias{getBestOutputPartitioning}
\alias{getBestOutputPartitioning.ddRDriver}
\title{This is an overrideable function that determines what the output
partitioning scheme of a dlapply or dmapply function should be.
It determines the 'ideal' nparts for the output if it is not supplied.
For API standard-enforcement, overriding this is not recommended.}
\usage{
getBestOutputPartitioning(driver, ..., nparts = NULL, type = NULL)

\method{getBestOutputPartitioning}{ddRDriver}(driver, ..., nparts = NULL,
  type = NULL)
}
\arguments{
\item{driver}{The backend driver to dispatch on.}

\item{...}{The arguments to this dmapply operation.}

\item{nparts}{The nparts argument, if any, supplied by the user.}

\item{type}{The output.type supplied by the user.}
}
\value{
A 2d-vector, that will be passed into your backend's do_dmapply.
}
\description{
This is an overrideable function that determines what the output
partitioning scheme of a dlapply or dmapply function should be.
It determines the 'ideal' nparts for the output if it is not supplied.
For API standard-enforcement, overriding this is not recommended.
}
\section{Methods (by class)}{
\itemize{
\item \code{ddRDriver}: The default implementation for
getBestOutputPartitioning.
}}

