% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getPartitionIdsAndOffsets}
\alias{getPartitionIdsAndOffsets}
\title{Gets the internal set of partitions, and offsets within each
partition, of a set 1d or 2d-subset indices for a distributed
object}
\usage{
getPartitionIdsAndOffsets(indices, psizes, nparts)
}
\arguments{
\item{indices}{A sorted list of sorted vectors, where the first element
are the row indices, and the second (if 2d), column indices.}

\item{psizes}{Partition-sizes matrix of the distributed object.}

\item{nparts}{nparts vector of the distributed object.}
}
\description{
It returns a list of 3 elements, where the first element is a list
of partitions, the second is a list of row indices, and third a
a list of column indices.
}
\details{
Note: This is an internal helper function of ddR.
}

