% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddR.R, R/pdriver.R
\docType{methods}
\name{init}
\alias{init}
\alias{init,ParallelddR-method}
\alias{init,ddRDriver-method}
\title{Called when the backend driver is initialized.}
\usage{
init(x, ...)

\S4method{init}{ddRDriver}(x, ...)

\S4method{init}{ParallelddR}(x, executors = NULL, type = "FORK", ...)
}
\arguments{
\item{x}{The driver object to initialize the backend for.}

\item{...}{Other parameters to pass to the initialization routine.}

\item{executors}{Number of cores to run with.}

\item{type}{If "FORK", will use UNIX fork() method. If "PSOCK", will use SNOW method.}
}
\description{
Called when the backend driver is initialized.
}
\section{Methods (by class)}{
\itemize{
\item \code{ddRDriver}: Default backend initialization message.

\item \code{ParallelddR}: Initialization for parallel
}}

