% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddR.R, R/pdriver.R
\docType{methods}
\name{init_driver}
\alias{init_driver}
\alias{init_driver,parallel.ddR-method}
\title{Initialize backend driver}
\usage{
init_driver(x, ...)

\S4method{init_driver}{parallel.ddR}(x, executors = "all",
  type = ifelse(windows, "PSOCK", "FORK"), ...)
}
\arguments{
\item{x}{The driver object to initialize the backend for.}

\item{...}{Other parameters to pass to the initialization routine.}

\item{executors}{Number of cores to run with, or "all" to use all
available cores}

\item{type}{If "FORK", will use UNIX fork() method. If "PSOCK", will use SNOW method.}

\item{...}{Additional arguments to \link[parallel]{makeCluster}}
}
\description{
Used internally by ddR to set up a new backend driver.
}
\section{Methods (by class)}{
\itemize{
\item \code{parallel.ddR}: Initialization for parallel
}}
\seealso{
\code{\link{useBackend}} for the user facing way to set or
     change backends
}

