% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdriver.R
\name{init_parallel}
\alias{init_parallel}
\title{Initialize the no. of cores in parallel backend}
\usage{
init_parallel(executors = "all", type = ifelse(windows, "PSOCK", "FORK"),
  ...)
}
\arguments{
\item{executors}{Number of cores to run with, or "all" to use all
available cores}

\item{type}{If "FORK", will use UNIX fork() method. If "PSOCK", will use SNOW method.}

\item{...}{Additional arguments to \link[parallel]{makeCluster}}
}
\value{
Object of class \code{\linkS4class{parallel.ddR}} representing a running parallel
     cluster
}
\description{
The FORK method of parallel works only on UNIX environments. The "PSOCK"
method requires SNOW but works on all OSes.
}

