% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{is.dframe}
\alias{is.DFrame}
\alias{is.dframe}
\title{Returns whether the input is a dframe}
\usage{
is.dframe(x)

is.DFrame(x)
}
\arguments{
\item{x}{input object.}
}
\value{
TRUE if x is a dframe, FALSE otherwise.
}
\description{
Returns whether the input is a dframe
}
\examples{
\dontrun{
is.dframe(3) # FALSE
is.dframe(dframe(psize=c(3,3),dim=c(9,9))) # TRUE
}
}

