% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dobject.R
\name{is.dobject}
\alias{is.DObject}
\alias{is.dobject}
\title{Returns whether the input entity is a DObject}
\usage{
is.dobject(x)

is.DObject(x)
}
\arguments{
\item{x}{The input to test to see whether it is a DObject.}
}
\value{
TRUE if x is a DObject, FALSE otherwise
}
\description{
Returns whether the input entity is a DObject
}
\examples{
\dontrun{
is.dobject(3) # FALSE
is.dobject(dlist(1,2,3,nparts=3)) # TRUE
is.dobject(darray(psize=c(3,3),dim=c(9,9))) # TRUE
}
}

