\name{major_scale}
\alias{major_scale}
\title{create a major scale given a base note}
\usage{
  major_scale(instrument = piano, root = "C3")
}
\arguments{
  \item{note}{the root note of the scale (eg "C3")}
}
\value{
  a data.frame of the notes in a scale
}
\description{
  create a major scale given a base note
}
\examples{
#not run
}

