\name{minor_scale}
\alias{minor_scale}
\title{select a major scale from an instrument given a root note}
\usage{
  minor_scale(instrument = piano, root = "C3")
}
\arguments{
  \item{instrument}{a built-in instrument object (eg
  "piano")}

  \item{note}{the root note of the scale (eg "C3")}
}
\value{
  a data.frame of the notes in a scale
}
\description{
  select a major scale from an instrument given a root note
}
\examples{
#not run
}

