% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{names<-,DObject-method}
\alias{names<-,DObject-method}
\title{Sets the names of a distributed object}
\usage{
\S4method{names}{DObject}(x) <- value
}
\arguments{
\item{x}{The object whose names to set.}

\item{value}{A vector with the names to set with.}
}
\description{
Sets the names of a distributed object
}

