% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdriver.R
\docType{class}
\name{parallel.ddR-class}
\alias{parallel.ddR-class}
\title{Class for parallel driver}
\arguments{
\item{type}{character "FORK" or "PSOCK"}
}
\description{
Class for parallel driver
}
\section{Slots}{

\describe{
\item{\code{cluster}}{As returned from \link[parallel]{makeCluster}}
}}

