% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdriver.R
\docType{data}
\name{parallel.ddR}
\alias{parallel.ddR}
\title{The default parallel driver}
\format{An object of class \code{parallel.ddR} of length 1.}
\usage{
parallel.ddR
}
\description{
The default parallel driver
}
\examples{
\dontrun{
useBackend(parallel.ddR, executors=4)
}
}
\keyword{datasets}

