\name{pitch}
\alias{pitch}
\title{change the pitch of a wav object.}
\usage{
  pitch(object, semitones, equal = TRUE)
}
\arguments{
  \item{object}{a wav object}

  \item{semitones}{the semitones with which to pitch the
  wav up(positive) or down(negative)}

  \item{equal}{whether the outuputted wave object should be
  the same length as the original}
}
\value{
  a repitched wave object
}
\description{
  change the pitch of a wav object.
}
\examples{
#not run
}

