% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{rbind,DObject-method}
\alias{rbind,DObject-method}
\title{row binds the arguments}
\usage{
\S4method{rbind}{DObject}(..., deparse.level = 1)
}
\arguments{
\item{...}{Arguments to row bind.}

\item{deparse.level}{Does nothing so far.}
}
\value{
A dobject with the opernads (and their partitions) rbinded.
}
\description{
row binds the arguments
}

