% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddR.R
\name{register_driver}
\alias{register_driver}
\title{Register a driver}
\usage{
register_driver(name, initfunc)
}
\arguments{
\item{name}{character, the common name for this backend. This will be
returned from \code{\link{available_backends}}}

\item{initfunc}{function capable of creating an instance of this driver
connected to a running backend.
Driver instances returned from this function should be a
subclass of \code{\linkS4class{ddRDriver}}}
}
\description{
Used internally to manage the different driver classes that ddR knows
about and supports. It should be used when creating new drivers for new backends.
}

