% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddR.R
\name{registerDriver}
\alias{registerDriver}
\title{Register a driver with ddR.}
\usage{
registerDriver(name, initfunc)
}
\arguments{
\item{name}{character, the common name for this backend. This will be
returned from \code{\link{availableBackend}}}

\item{initfunc}{function capable of creating an instance of this new
driver. Objects returned from this function should be a 
subclass of \code{\link{ddRDriver}}}
}
\description{
This function is used internally to manage the drivers that ddR knows
about and supports.
}

