% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{rowMeans,DObject-method}
\alias{rowMeans,DObject-method}
\title{Gets the row means for a distributed array or data.frame.}
\usage{
\S4method{rowMeans}{DObject}(x, na.rm = FALSE, dims = 1L)
}
\arguments{
\item{x}{The object to get the row means from.}

\item{na.rm}{If TRUE, will remove NAs.}

\item{dims}{Currently does nothing.}
}
\description{
Gets the row means for a distributed array or data.frame.
}

