% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddR.R
\name{shutdown}
\alias{shutdown}
\title{Called when the backend driver is shutdown.}
\usage{
shutdown(x)
}
\arguments{
\item{x}{The driver object to shutdown, defaults to the current one.}
}
\description{
Called when the backend driver is shutdown.
}

