% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{[[,DObject,numeric-method}
\alias{[[,DObject,numeric-method}
\title{Extracts a single element of a distributed object.}
\usage{
\S4method{[[}{DObject,numeric}(x, i, j, ...)
}
\arguments{
\item{x}{The object to get an element from.}

\item{i}{The row index of the element.}

\item{j}{The column index of the element.}

\item{...}{Other args}
}
\description{
Extracts a single element of a distributed object.
}

