% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\docType{methods}
\name{sum,DObject-method}
\alias{sum,DObject-method}
\title{Gets the sum of the objects.}
\usage{
\S4method{sum}{DObject}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{The first distributed object}

\item{...}{Other objects}

\item{na.rm}{If TRUE, removes the NA values.}
}
\description{
Gets the sum of the objects.
}

