% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddR.R
\name{validate_dargs}
\alias{validate_dargs}
\title{Validate the arguments for dmapply}
\usage{
validate_dargs(...)
}
\arguments{
\item{...}{iterable arguments for dmapply}
}
\value{
list of validated arguments
}
\description{
Validate the arguments for dmapply
}

