\name{combine}
\alias{combine}
\title{combine multiple waves objects together}
\usage{
  combine(objects, unit = "16")
}
\arguments{
  \item{objects}{a list of wave objects of the same length
  to be combined}

  \item{the}{bit rate of the objects}
}
\value{
  a single wave object
}
\description{
  combine multiple waves objects together
}
\examples{
#not run
}

