% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddR.R
\docType{class}
\name{ddRDriver-class}
\alias{ddRDriver-class}
\title{The base S4 class for backend driver classes to extend.}
\description{
The base S4 class for backend driver classes to extend.
}
\section{Slots}{

\describe{
\item{\code{DListClass}}{A character vector naming the class-name for dlists.}

\item{\code{DArrayClass}}{A character vector naming the class-name for darrays.}

\item{\code{DFrameClass}}{A character vector naming the class-name for dframes.}

\item{\code{backendName}}{A character vector naming the backend.}
}}

