% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddR.R
\name{registerDriver}
\alias{registerDriver}
\title{Register a driver with ddR.}
\usage{
registerDriver(name, initfunc)
}
\description{
This function is used internally to manage the drivers that ddR knows
about and supports. It's purpose is to 
a new driver for ddR.
}

