% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddR.R, R/pdriver.R
\docType{methods}
\name{shutdown}
\alias{shutdown}
\alias{shutdown,ParallelddR-method}
\alias{shutdown,ddRDriver-method}
\title{Called when the backend driver is shutdown.}
\usage{
shutdown(x)

\S4method{shutdown}{ddRDriver}(x)

\S4method{shutdown}{ParallelddR}(x)
}
\arguments{
\item{x}{The driver object to shutdown.}
}
\description{
Called when the backend driver is shutdown.
}
\section{Methods (by class)}{
\itemize{
\item \code{ddRDriver}: Default backend shutdown message.

\item \code{ParallelddR}: Shutdown for parallel
}}

