% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddR.R, R/pdriver.R
\docType{methods}
\name{shutdown}
\alias{shutdown}
\alias{shutdown,parallel.ddR-method}
\title{Called when the backend driver is shutdown.}
\usage{
shutdown(x)

\S4method{shutdown}{parallel.ddR}(x)
}
\arguments{
\item{x}{The driver object to shutdown, defaults to the current one.}
}
\description{
Called when the backend driver is shutdown.
}
\section{Methods (by class)}{
\itemize{
\item \code{parallel.ddR}: Shutdown for parallel
}}

